Refactoring: Separate files
