+++ money/interfaces.go
@@ -1,0 +1,7 @@
+package money
+
+type Expression interface {
+	reduce(b *Bank, currency string) *Money
+	Plus(addend Expression) Expression
+	Times(multiplier int) Expression
+}
